Website_Auditor/requirements.txt
gradio>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
openai>=1.3.5
python-dotenv>=1.0.0
